Build out the CI check suite
Six workflows, all of them gates the tree already passes:

ci          gcc, clang, macOS and arm64 builds with -Werror, plus Debug; the
            ctest suite; AddressSanitizer with UndefinedBehaviorSanitizer;
            ThreadSanitizer, which is the only check that looks at the
            pthreads solvers and their atomics; valgrind memcheck on the ten
            fast suites; an install-and-consume job that builds a real
            consumer through find_package and then again through pkg-config
            from a moved prefix, which is the only test of the relocatable
            .pc file; and the CUDA kernel compiled for sm_70 to sm_90.
analysis    clang-tidy, cppcheck, gcc -fanalyzer, clang-format over the lines
            a change touches, shellcheck, actionlint, and gcovr with a floor
            under the current 91% lines and 92% functions.
codeql      C, Go and Python with the security-and-quality pack.
nightly     valgrind on the two suites that take an hour each, every
            benchmark under both sanitizer sets, a recorded benchmark run,
            and a wider OS matrix.

Formatting is enforced on changed lines rather than the whole tree: the
sources predate .clang-format and reformatting them would bury every future
diff behind 65 files of whitespace.  New and edited code still has to match.

CA_SANITIZE now takes a sanitizer list, so the same switch selects
address,undefined or thread; ON still means address,undefined.  Every
executable picks the flags up through one helper, since linking a sanitized
library into an unsanitized binary leaves the runtime's symbols undefined.

The Makefile gained tidy, cppcheck, analyzer, shellcheck, format, tsan,
valgrind and coverage targets, and a checks target that runs the
pull-request set in the order that fails fastest.

Also adds dependabot for actions, cargo, go and pip; CODEOWNERS; a pull
request template that asks for the before/after constants on an algorithmic
change; and two issue forms.

Four tests needed casts or an added free to pass clang-tidy on tests/, which
the clang-tidy job now covers.

+++ tests/test_linalg.c
@@ -35,7 +35,7 @@ int main(void)
     {
         uint64_t q = 1000003;
         uint32_t n = 40;
-        uint64_t *M = calloc(n * n, sizeof(uint64_t)), *rhs = calloc(n, sizeof(uint64_t));
+        uint64_t *M = calloc((size_t)n * n, sizeof(uint64_t)), *rhs = calloc(n, sizeof(uint64_t));
         uint64_t *xt = calloc(n, sizeof(uint64_t)), *x = calloc(n, sizeof(uint64_t));
         for (uint32_t j = 0; j < n; j++) xt[j] = ca_rng_below(&rng, q);
         for (uint32_t i = 0; i < n; i++) {
